Fractalizer.jl documentation

Fractalizer.fractalize — Function

fractalize(shape::T, template::Template) where where T <: AbstractShape
fractalize(shape::T, noise_params::NoiseParams) where where T <: AbstractShape
fractalize(shape::T, template::Template, iter::Int) where where T <: AbstractShape
fractalize(shape::T, noise_params::NoiseParams, iter::Int) where where T <: AbstractShape

Fractalize an AbstractShape using either a template or random noise with additional cumulative steps.

Input

shape – AbstractShape to fractalize

template – Template used to fractalize

noise_params – Settings for the random noise

iter – Number of times to apply the fractalization algorithm

Output

A new AbstractShape of the same concrete type of shape.

Algorithm

The algorithm works by scaling and rotating the template for each segment of the shape. The new shape will result in the combination of the scaled and rotated templates at each segment. This process can be repeated multiple times to get a fractal where each segment of the new shape will be modified following the template. To scale we calculate the length of the segment and the line_length of the template and scale the template depending on this value. The translation of the template is done so that the first point of the template coincides with the first point of the segment.

When using the noise_params each segment will have a unique noise signal i.e. each segment will have a unique template.

Fractalizer.Shape — Type

Shape <: AbstractShape

Type to encapsulate shapes that need to be fractalized.

Fields

centroid – [x, y] vector of the coordinates of the centroid of the shape

npoints – Number of points of the shape

nsegments – Number of segments of the shape

segment_lengths – Lengths of each segment of the shape

segment_angles – Azimuthal angle (0,360) of each segment in the shape, with [1,0] as reference 0°.

segment_centers – Center of each segment of the shape.

segment_normals – Normals of each segment of the shape.

points – nx2 Matrix of points

edges – nx2 connectivity Matrix indicating the indexes of the points that define the segment

bb – 4x2 Matrix indicating the bounding box of the shape. [[minimum(xs), maximum(xs)] [minimum(ys), maximum(ys)]]

xs – x coordinates of the points

ys – y coordinates of the points

l – bounding box length

w – bounding box width

Notes

xs and ys are views of points

Examples

Shape(centroid::Vector{Float64}, npoints::Int, nsegments::Int segment_lengths::Vector{Float64}, segment_angles::Vector{Float64}, segment_centers::Matrix{Float64}, segment_normals::Matrix{Float64}, points::Matrix{Float64}, edges::Matrix{Int}, bb::Matrix{Float64}, xs::SubArray, ys::SubArray, l::Float64, w::Float64) – default constructor

Shape(points::Matrix{Float64}) – most used constructor

Fractalizer.ClosedShape — Type

ClosedShape <: AbstractShape

Type to encapsulate closed shapes that need to be fractalized (i.e. rings or closed polygons).

Fields

centroid – [x, y] vector of the coordinates of the centroid of the shape

npoints – Number of points of the shape

nsegments – Number of segments of the shape

segment_lengths – Lengths of each segment of the shape

segment_angles – Azimuthal angle (0,360) of each segment in the shape, with [1,0] as reference 0°.

segment_centers – Center of each segment of the shape.

segment_normals – Normals of each segment of the shape.

points – nx2 Matrix of points

edges – nx2 connectivity Matrix indicating the indexes of the points that define the segment

bb – 4x2 Matrix indicating the bounding box of the shape. [[minimum(xs), maximum(xs)] [minimum(ys), maximum(ys)]]

xs – x coordinates of the points

ys – y coordinates of the points

l – bounding box length

w – bounding box width

Notes

xs and ys are views of points

The last and first points of the shape coincide. This is automatically done when inputting a matrix of points.

Examples

ClosedShape(centroid::Vector{Float64}, npoints::Int, nsegments::Int segment_lengths::Vector{Float64}, segment_angles::Vector{Float64}, segment_centers::Matrix{Float64}, segment_normals::Matrix{Float64}, points::Matrix{Float64}, edges::Matrix{Int}, bb::Matrix{Float64}, xs::SubArray, ys::SubArray, l::Float64, w::Float64) – default constructor

ClosedShape(points::Matrix{Float64}) – most used constructor

Fractalizer.Template — Type

Template <: AbstractTemplate

Type to encapsulate templates used to fractalize.

Fields

centroid – [x, y] vector of the coordinates of the centroid of the template

npoints – Number of points of the template

line_length – Length of the template as distance between the first and last point

line_angle – Inclination of the template using the first and last point.

points – nx2 Matrix of points

xs – x coordinates of the points

ys – y coordinates of the points

centered – CenteredTemplate copy of the template. This is used to properly scale the template that then is translated to the segment

Notes

xs and ys are views of points

Examples

Template(centroid::Vector{Float64}, npoints::Int, line_length::Float64, line_angle::Float64, points::Matrix{Float64}, xs::SubArray, ys::SubArray, centered::CenteredTemplate) – default constructor

Template(points::Matrix{Float64}) – most used constructor

Fractalizer.NoiseParams — Type

NoiseParams

Type used to encapsulate the random settings for generating random templates.

Fields

amplitude_range – Range of amplitudes to be used

frequency_range – Range of frequency to be used

phase_range – Range of phases to be used

resolution – N of points of the generated random signal

iterations – N of times that random signals are stacked

nsamples – N of samples of the final random template

seed – seed for random generation

Examples

NoiseParams(amplitude_range::AbstractRange, frequency_range::AbstractRange, phase_range::AbstractRange, resolution::Int, iterations::Int, nsamples::Int, seed::Int) – default constructor

Fractalizer.random_template — Function

random_template(params::NoiseParams)

Generate a random template given the parameters in NoiseParams

Input

params – NoiseParams struct with all the settings for random generation

Output

A Template object

Algorithm

The random template is defined as follows:

A random value of amplitude, frequency and phase are picked from the ranges

A sine wave in (0,2π) is generated with the picked resolution and the random values

A random selection of nsamples are then picked from the signal

Fractalizer.makering — Function

makering(x0, y0, r, npoints)

Create a ring.

Input

x0 – x center of the ring

y0 – y center of the ring

r – radius of the ring

npoints – n of points of the ring

Output

A ClosedShape object

Notes

Since the output is a closed shape, the number of sides is 1-npoints. This means that

triangle -> npoints = 4

square -> npoints = 5

pentagon -> npoints = 6

...

Fractalizer.R — Function

R(θ)

2D Rotation matrix for angle θ in degrees.

$[[cosd(θ), sind(θ)] [-sind(θ), cosd(θ)]]$

Input

θ – Rotation angle in degrees.

Output

2x2 rotation matrix