Functions

Fractalizer.fractalize — Function

fractalize(shape::T, template::Template) where T <: AbstractShape
fractalize(shape::T, noise_params::NoiseParams) where T <: AbstractShape
fractalize(shape::T, template::Template, iter::Int) where T <: AbstractShape
fractalize(shape::T, noise_params::NoiseParams, iter::Int) where T <: AbstractShape

Fractalize an AbstractShape using either a template or random noise with additional cumulative steps.

Input

shape – AbstractShape to fractalize

template – Template used to fractalize

noise_params – Settings for the random noise

iter – Number of times to apply the fractalization algorithm

Output

A new AbstractShape of the same concrete type of shape.

Algorithm

The algorithm works by scaling and rotating the template for each segment of the shape. The new shape will result in the combination of the scaled and rotated templates at each segment. This process can be repeated multiple times to get a fractal where each segment of the new shape will be modified following the template. To scale we calculate the length of the segment and the line_length of the template and scale the template depending on this value. The translation of the template is done so that the first point of the template coincides with the first point of the segment.

When using the noise_params each segment will have a unique noise signal i.e. each segment will have a unique template.

Fractalizer.random_template — Function

random_template(params::NoiseParams)

Generate a random template given the parameters in NoiseParams

Input

params – NoiseParams struct with all the settings for random generation

Output

A Template object

Algorithm

The random template is defined as follows:

A random value of amplitude, frequency and phase are picked from the ranges

A sine wave in (0,2π) is generated with the picked resolution and the random values

A random selection of nsamples are then picked from the signal

Fractalizer.makering — Function

makering(x0, y0, r, npoints)

Create a ring.

Input

x0 – x center of the ring

y0 – y center of the ring

r – radius of the ring

npoints – n of points of the ring

Output

A ClosedShape object

Notes

Since the output is a closed shape, the number of sides is 1-npoints. This means that

triangle -> npoints = 4

square -> npoints = 5

pentagon -> npoints = 6

...

Fractalizer.R — Function

R(θ)

2D Rotation matrix for angle θ in degrees.

\[R(\theta) = \begin{bmatrix} \cos{\theta} & \sin{\theta}\\ -\sin{\theta} & \cos{\theta}\\ \end{bmatrix}\]

Input

θ – Rotation angle in degrees.

Output

2x2 rotation matrix