Types

Fractalizer.Shape — Type

Shape <: AbstractShape

Type to encapsulate shapes that need to be fractalized.

Fields

centroid::Vector{Float64} – [x, y] vector of the coordinates of the centroid of the shape

npoints::Int – Number of points of the shape

nsegments::Int – Number of segments of the shape

segment_lengths::Vector{Float64} – Lengths of each segment of the shape

segment_angles::Vector{Float64} – Azimuthal angle (0,360) of each segment in the shape, with [1,0] as reference 0°.

segment_centers::Matrix{Float64} – Center of each segment of the shape.

segment_normals::Matrix{Float64} – Normals of each segment of the shape.

points::Matrix{Float64} – nx2 Matrix of points

edges::Matrix{Int} – nx2 connectivity Matrix indicating the indexes of the points that define the segment

bb::Matrix{Float64} – 4x2 Matrix indicating the bounding box of the shape. [[minimum(xs), maximum(xs)] [minimum(ys), maximum(ys)]]

xs::SubArray – x coordinates of the points

ys::SubArray – y coordinates of the points

l::Float64 – bounding box length

w::Float64 – bounding box width

Notes

xs and ys are views of points

Examples

Shape(points::Matrix{Float64}) – most used constructor

Fractalizer.ClosedShape — Type

ClosedShape <: AbstractShape

Type to encapsulate closed shapes that need to be fractalized (i.e. rings or closed polygons).

Fields

centroid::Vector{Float64} – [x, y] vector of the coordinates of the centroid of the shape

npoints::Int – Number of points of the shape

nsegments::Int – Number of segments of the shape

segment_lengths::Vector{Float64} – Lengths of each segment of the shape

segment_angles::Vector{Float64} – Azimuthal angle (0,360) of each segment in the shape, with [1,0] as reference 0°.

segment_centers::Matrix{Float64} – Center of each segment of the shape.

segment_normals::Matrix{Float64} – Normals of each segment of the shape.

points::Matrix{Float64} – nx2 Matrix of points

edges::Matrix{Int} – nx2 connectivity Matrix indicating the indexes of the points that define the segment

bb::Matrix{Float64} – 4x2 Matrix indicating the bounding box of the shape. [[minimum(xs), maximum(xs)] [minimum(ys), maximum(ys)]]

xs::SubArray – x coordinates of the points

ys::SubArray – y coordinates of the points

l::Float64 – bounding box length

w::Float64 – bounding box width

Notes

xs and ys are views of points

The last and first points of the shape coincide. This is automatically done when inputting a matrix of points.

Examples

ClosedShape(points::Matrix{Float64}) – most used constructor

Fractalizer.Template — Type

Template <: AbstractTemplate

Type to encapsulate templates used to fractalize.

Fields

centroid::Vector{Float64} – [x, y] vector of the coordinates of the centroid of the template

npoints::Int – Number of points of the template

line_length::Float64 – Length of the template as distance between the first and last point

line_angle::Float64 – Inclination of the template using the first and last point.

points::Matrix{Float64} – nx2 Matrix of points

xs::SubArray – x coordinates of the points

ys::SubArray – y coordinates of the points

centered::CenteredTemplate – CenteredTemplate copy of the template. This is used to properly scale the template that then is translated to the segment

Notes

xs and ys are views of points

Examples

Template(points::Matrix{Float64}) – most used constructor

Fractalizer.NoiseParams — Type

NoiseParams

Type used to encapsulate the random settings for generating random templates.

Fields

amplitude_range::AbstractRange – Range of amplitudes to be used

frequency_range::AbstractRange – Range of frequency to be used

phase_range::AbstractRange – Range of phases to be used

resolution::Int – N of points of the generated random signal

iterations::Int – N of times that random signals are stacked

nsamples::Int – N of samples of the final random template

seed::Int – seed for random generation

Examples

noise_params = NoiseParams(0.1:0.1, 1.0:1:10.0, -10.0:1:10.0, 100, 4, 10) – default constructor, random seed

noise_params = NoiseParams(0.1:0.1, 1.0:1:10.0, -10.0:1:10.0, 100, 4, 10, seed=1) – default constructor, fixed seed